refactor: correção de warnings em clientes

+++ cliente/clientes.c
@@ -213,7 +213,7 @@ void atualizarCliente(void) {
             strcpy(c.telefone, tel);
             free(tel);
 
-            fseek(fp, -sizeof(Cliente), SEEK_CUR);
+            fseek(fp, -(long) sizeof(Cliente), SEEK_CUR);
             fwrite(&c, sizeof(Cliente), 1, fp);
             encontrado = 1;
             break;
@@ -244,7 +244,7 @@ void excluirCliente(void) {
     while (fread(&c, sizeof(Cliente), 1, fp) == 1) {
         if (strcmp(c.id, id) == 0 && c.status == 1) {
             c.status = 0;
-            fseek(fp, -sizeof(Cliente), SEEK_CUR);
+            fseek(fp, -(long) sizeof(Cliente), SEEK_CUR);
             fwrite(&c, sizeof(Cliente), 1, fp);
             encontrado = 1;
             break;
